requirements.txt
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.26.0
aiohttp>=3.9.0
scipy>=1.11.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Phase 4 (trading execution) — requires Python >=3.9.10
# py-clob-client>=0.0.1
# eth-account>=0.11.0
# web3>=6.0.0
